# docs/openapi/tracebud-v1-draft.yaml
openapi: 3.0.3
info:
  title: Tracebud API (v1 Draft)
  version: "0.1.0-draft"
  description: |
    Draft OpenAPI artifact generated from TRACEBUD_V1_2_EUDR_SPEC.md Section 32.8.
    This document is intentionally minimal and should be expanded during implementation.
  license:
    name: Proprietary
    url: https://tracebud.com/legal
  x-tracebud-error-mapping:
    AUTH-001: INVALID_TOKEN
    AUTH-002: INSUFFICIENT_SCOPE
    VAL-001: FIELD_REQUIRED
    VAL-002: FIELD_INVALID
    TEN-001: TENANT_SCOPE_VIOLATION
    CON-001: CONSENT_REQUIRED
    SYNC-001: IDEMPOTENCY_REPLAY
    SYNC-002: CONFLICT_REVIEW_REQUIRED
    GEO-001: GEOMETRY_INVALID
    GEO-002: GEOMETRY_VARIANCE_EXCEEDED
    YLD-001: BENCHMARK_UNAVAILABLE
    DDS-001: INVALID_STATE_TRANSITION
    DDS-002: SUBMISSION_REJECTED
    EXT-001: TRACES_AUTH_FAILED
    EXT-002: TRACES_VALIDATION_FAULT
    BIL-001: PAYMENT_FAILED
    OPS-001: INTERNAL_ERROR
servers:
  - url: https://api.tracebud.com
    description: Production API
  - url: https://sandbox-api.tracebud.com
    description: Sandbox API
security:
  - bearerAuth: []
...
paths:
...
  /v1/integrations/eudr/echo:
    get:
      tags: [Integrations]
      operationId: checkEudrConnectivity
      summary: Check EUDR connectivity
      description: Performs server-side connectivity check against EUDR echo endpoint using backend-managed credentials.
      parameters:
        - name: message
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: EUDR connectivity check succeeded
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - provider
                  - endpoint
                  - statusCode
                  - version
                  - latencyMs
                  - message
                properties:
                  ok:
                    type: boolean
                  provider:
                    type: string
                    example: eudr_api_service
                  endpoint:
                    type: string
                    example: /api/eudr/echo
                  statusCode:
                    type: integer
                  version:
                    type: string
                    example: "2"
                  latencyMs:
                    type: integer
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Missing tenant claim or role policy violation
        "502":
          description: Upstream EUDR connectivity check failed
  /v1/integrations/eudr/dds:
    post:
      tags: [Integrations]
      operationId: submitEudrDds
      summary: Submit DDS to EUDR
      description: Submits tenant-scoped DDS payload to EUDR provider and returns replay-aware outcome metadata.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - statement
                - idempotencyKey
              properties:
                statement:
                  type: object
                  additionalProperties: true
                idempotencyKey:
                  type: string
      responses:
        "200":
          description: DDS submit accepted (including replay-safe duplicate outcomes)
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - provider
                  - endpoint
                  - idempotencyKey
                  - statusCode
                  - version
                  - latencyMs
                properties:
                  ok:
                    type: boolean
                  provider:
                    type: string
                    example: eudr_api_service
                  endpoint:
                    type: string
                    example: /api/eudr/dds
                  idempotencyKey:
                    type: string
                  statusCode:
                    type: integer
                  version:
                    type: string
                    example: "2"
                  latencyMs:
                    type: integer
                  replayed:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Missing tenant claim or role policy violation
        "502":
          description: Upstream EUDR DDS submit failed
  /v1/integrations/eudr/dds/status:
    get:
      tags: [Integrations]
      operationId: readEudrDdsStatus
      summary: Read DDS status from EUDR
      parameters:
        - name: referenceNumber
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: EUDR DDS status read succeeded
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - provider
                  - endpoint
                  - referenceNumber
                  - statusCode
                  - version
                  - latencyMs
                properties:
                  ok:
                    type: boolean
                  provider:
                    type: string
                    example: eudr_api_service
                  endpoint:
                    type: string
                    example: /api/eudr/dds
                  referenceNumber:
                    type: string
                  statusCode:
                    type: integer
                  version:
                    type: string
                    example: "2"
                  latencyMs:
                    type: integer
                  payload:
                    type: object
                    additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Missing tenant claim or role policy violation
        "502":
          description: Upstream EUDR DDS status read failed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error_code: { type: string }
        message: { type: string }
        field_errors:
          type: array
          items: { type: object }
        request_id: { type: string }
      required: [error_code, message, request_id]
...